+++ src/classes/timelineWords4.ts
@@ -1,0 +1,255 @@
+import { enemy } from 'enemies';
+import { mough, pouler, shiden } from 'friends';
+import { Timelines } from './Timelines';
+import { Direction } from './Direction';
+import { castleAnnounce, oceanGotsuji, volcanoGotsuji } from './timelineWords0';
+//Stage4
+export const explanation4: Timelines = {
+  start: [
+    {
+      type: 'dialog',
+      text: 'ここはキーマボルケーノ。',
+      actorName: '杜松の看板',
+    },
+    {
+      type: 'dialog',
+      text: '溶岩石で出来た自然の迷宮。',
+      actorName: '杜松の看板',
+    },
+    { type: 'endTimeline' },
+  ],
+};
+
+export const addMough: Timelines = {
+  start: [
+    { type: 'event', event: 'set', props: ['Shiden', 3, 2, undefined] },
+    { type: 'event', event: 'chdir', props: ['Shiden', 'down'] },
+    { type: 'event', event: 'set', props: ['Pouler', 2, 1, undefined] },
+    { type: 'event', event: 'chdir', props: ['Pouler', 'down'] },
+    { type: 'event', event: 'set', props: ['Mough', 2, 1, undefined] },
+    { type: 'event', event: 'chdir', props: ['Mough', 'down'] },
+    { type: 'event', event: 'log', props: ['player', 1] },
+    { type: 'dialog', actorName: 'マルク', text: 'なぁ…なんか、着いてきてね？' },
+    { type: 'event', event: 'log', props: ['Shiden', 1] },
+    { type: 'dialog', actorName: 'シデン', text: 'なんか、ではない。' },
+    {
+      type: 'dialog',
+      actorName: 'シデン',
+      text: '逆にこんな大男が着いてきていてなぜ今まで気付かなかった。',
+    },
+    { type: 'event', event: 'log', props: ['Shiden', 3] },
+    { type: 'dialog', actorName: 'シデン', text: 'なんの用だ、大男。' },
+    { type: 'event', event: 'log', props: ['Mough', 3] },
+    { type: 'dialog', actorName: 'マウ', text: '…おで？' },
+    { type: 'event', event: 'log', props: ['Shiden', 2] },
+    { type: 'dialog', actorName: 'シデン', text: '貴様以外に誰がいる。' },
+    { type: 'event', event: 'relog', props: [] },
+    { type: 'dialog', actorName: 'マウ', text: 'お前ら、食堂、行く。' },
+    { type: 'dialog', actorName: 'マルク', text: 'そうだな。' },
+    { type: 'dialog', actorName: 'マウ', text: '食堂、魔王、居る。' },
+    { type: 'dialog', actorName: 'パウラ', text: 'そうですね。' },
+    { type: 'dialog', actorName: 'マウ', text: '魔王、強い。' },
+    { type: 'dialog', actorName: 'シデン', text: 'その通りだ。' },
+    { type: 'event', event: 'log', props: ['Mough', 3] },
+    { type: 'dialog', actorName: 'マウ', text: 'マウも、強い。' },
+    { type: 'event', event: 'log', props: ['Shiden', 2] },
+    { type: 'dialog', actorName: 'シデン', text: 'それは知らん。' },
+    { type: 'event', event: 'log', props: ['Mough', 5] },
+    { type: 'dialog', actorName: 'マウ', text: '腕試し、する。' },
+    { type: 'event', event: 'log', props: ['player', 4] },
+    { type: 'dialog', actorName: 'マルク', text: '…それで着いてきてるのか。' },
+    { type: 'event', event: 'relog', props: [] },
+    { type: 'dialog', actorName: 'マルク', text: 'まぁ、力強そうだし。' },
+    { type: 'event', event: 'log', props: ['player', 6] },
+    { type: 'dialog', actorName: 'マルク', text: 'いいんじゃない？' },
+    { type: 'event', event: 'log', props: ['Shiden', 5] },
+    { type: 'dialog', actorName: 'シデン', text: 'いいな。' },
+    { type: 'event', event: 'log', props: ['Pouler', 3] },
+    { type: 'dialog', actorName: 'パウラ', text: 'いいんですか！？' },
+    { type: 'event', event: 'log', props: ['Mough', 5] },
+    { type: 'dialog', actorName: 'マウ', text: 'マウ、強い、守る。' },
+    { type: 'event', event: 'log', props: ['Pouler', 4] },
+    { type: 'dialog', actorName: 'パウラ', text: '確かに頼もしくはありますけど…。' },
+    { type: 'setBackgroundColor', color: '#000' },
+    { type: 'event', event: 'reset', props: ['Shiden'] },
+    { type: 'event', event: 'reset', props: ['Pouler'] },
+    { type: 'event', event: 'reset', props: ['Mough'] },
+    { type: 'event', event: 'delete', props: ['add4'] },
+    { type: 'event', event: 'relog', props: [] },
+    { type: 'dialog', text: 'ノリと勢いでマウが仲間になった！' },
+    { type: 'meetFriend', actor: mough },
+    { type: 'endTimeline' },
+  ],
+};
+
+export const warpboss: Timelines = {
+  start: [
+    { type: 'dialog', text: 'この穴に入りますか？' },
+    {
+      type: 'choice',
+      choices: [
+        { text: 'はい', timelineID: 'Yes' },
+        { text: 'いいえ', timelineID: 'No' },
+      ],
+    },
+  ],
+  Yes: [{ type: 'event', event: 'warp', props: [29, 45] }, { type: 'endTimeline' }],
+  No: [{ type: 'endTimeline' }],
+};
+export const backboss: Timelines = {
+  start: [
+    { type: 'dialog', text: 'この穴に入りますか？' },
+    {
+      type: 'choice',
+      choices: [
+        { text: 'はい', timelineID: 'Yes' },
+        { text: 'いいえ', timelineID: 'No' },
+      ],
+    },
+  ],
+  Yes: [{ type: 'event', event: 'warp', props: [4, 29] }, { type: 'endTimeline' }],
+  No: [{ type: 'endTimeline' }],
+};
+export const goEleca: Timelines = {
+  start: [
+    { type: 'setBackgroundColor', color: '#000' },
+    { type: 'event', event: 'warp', props: [29, 43] },
+    { type: 'event', event: 'set', props: ['Shiden', 28, 43, undefined] },
+    { type: 'event', event: 'set', props: ['Pouler', 30, 44, undefined] },
+    { type: 'event', event: 'set', props: ['Mough', 27, 44, undefined] },
+    { type: 'event', event: 'chdir', props: ['player', 'up'] },
+    { type: 'event', event: 'chdir', props: ['Shiden', 'up'] },
+    { type: 'event', event: 'chdir', props: ['Pouler', 'up'] },
+    { type: 'event', event: 'chdir', props: ['Mough', 'up'] },
+    { type: 'endTimeline' },
+  ],
+};
+export const beforeElecaBattle: Timelines = {
+  start: [
+    { type: 'event', event: 'log', props: ['player', 5] },
+    {
+      type: 'dialog',
+      actorName: 'マルク',
+      text: 'どうせここの四天王も、威勢だけなんだろうなぁ。',
+    },
+    { type: 'event', event: 'log', props: ['Mough', 3] },
+    { type: 'dialog', actorName: 'マウ', text: '待て。' },
+    { type: 'event', event: 'chdir', props: ['Pouler', 'left'] },
+    { type: 'event', event: 'log', props: ['Pouler', 1] },
+    { type: 'dialog', actorName: 'パウラ', text: 'どうしたんですか？マウさん。' },
+    { type: 'event', event: 'log', props: ['Mough', 4] },
+    { type: 'dialog', actorName: 'マウ', text: 'こいつ、強い…。' },
+    { type: 'event', event: 'bosslog', props: [5] },
+    { type: 'event', event: 'chdir', props: ['Pouler', 'up'] },
+    {
+      type: 'dialog',
+      actorName: 'エレカ',
+      text: 'あら、なかなか見る目がある仲間を連れてるじゃない。',
+    },
+    { type: 'event', event: 'log', props: ['Shiden', 3] },
+    {
+      type: 'dialog',
+      actorName: 'シデン',
+      text: 'なんだこいつ…今までのとは気迫が違う…。',
+    },
+    { type: 'event', event: 'log', props: ['player', 1] },
+    {
+      type: 'dialog',
+      actorName: 'マルク',
+      text: '…あれ？もしかしてピンと来てないの俺だけ？',
+    },
+    { type: 'event', event: 'relog', props: [] },
+    {
+      type: 'dialog',
+      actorName: 'エレカ',
+      text: 'あの3人と同じだと思ってかかると痛い目に合うわよ。',
+    },
+    { type: 'event', event: 'bosslog', props: [5] },
+    {
+      type: 'dialog',
+      actorName: 'エレカ',
+      text: 'ま、ここに足を踏み入れた時点で命は無いと思うことね。',
+    },
+    { type: 'event', event: 'log', props: ['player', 3] },
+    {
+      type: 'dialog',
+      actorName: 'マルク',
+      text: '言ってくれるじゃねぇか、こっちは4人だぜ。',
+    },
+    { type: 'event', event: 'bosslog', props: [4] },
+    {
+      type: 'dialog',
+      actorName: 'エレカ',
+      text: '何人いたって関係ないわ。',
+    },
+    { type: 'event', event: 'bosslog', props: [3] },
+    {
+      type: 'dialog',
+      actorName: 'エレカ',
+      text: '全員まとめて叩き潰す。',
+    },
+    { type: 'event', event: 'relog', props: [] },
+    { type: 'endTimeline' },
+  ],
+};
+export const afterElecaBattle: Timelines = {
+  start: [
+    { type: 'event', event: 'bosslog', props: [4] },
+    { type: 'dialog', actorName: 'エレカ', text: 'うぐっ…おかあ…さま…。' },
+    { type: 'event', event: 'log', props: ['player', 4] },
+    { type: 'dialog', actorName: 'マルク', text: 'はぁ…はぁ…流石に歯応えあったな…。' },
+    { type: 'event', event: 'bosslog', props: [4] },
+    {
+      type: 'dialog',
+      actorName: 'エレカ',
+      text: '…お母様は、私の比にならない強さよ、\nせいぜい…覚悟…しておきなさい…。',
+    },
+    { type: 'event', event: 'relog', props: [] },
+    { type: 'event', event: 'break', props: ['Eleca'] },
+    { type: 'dialog', actorName: 'エレカ', text: 'バタッ。' },
+    { type: 'event', event: 'chdir', props: ['player', 'down'] },
+    { type: 'event', event: 'chdir', props: ['Shiden', 'down'] },
+    { type: 'event', event: 'chdir', props: ['Pouler', 'left'] },
+    { type: 'event', event: 'chdir', props: ['Mough', 'right'] },
+    { type: 'event', event: 'log', props: ['Pouler', 1] },
+    { type: 'dialog', actorName: 'パウラ', text: '皆さん、お怪我は大丈夫ですか…？' },
+    { type: 'event', event: 'log', props: ['Mough', 5] },
+    { type: 'dialog', actorName: 'マウ', text: 'マウ、丈夫、マウ、平気。' },
+    { type: 'event', event: 'relog', props: [] },
+    { type: 'dialog', actorName: 'パウラ', text: '良かったです…。' },
+    { type: 'event', event: 'log', props: ['Shiden', 4] },
+    {
+      type: 'dialog',
+      actorName: 'シデン',
+      text: 'あいつの話によれば、OBCはこれの比にならない強さらしいな。',
+    },
+    { type: 'event', event: 'log', props: ['player', 1] },
+    {
+      type: 'dialog',
+      actorName: 'マルク',
+      text: 'ああ、折角ここまで来たんだ。\nもっと強くなって、魔王ぶっとばしてやろうぜ！',
+    },
+    { type: 'event', event: 'relog', props: [] },
+    { type: 'setBackgroundColor', color: '#000' },
+    { type: 'event', event: 'zoomDown', props: [] },
+    { type: 'event', event: 'reset', props: ['Shiden'] },
+    { type: 'event', event: 'reset', props: ['Pouler'] },
+    { type: 'event', event: 'reset', props: ['Mough'] },
+    { type: 'event', event: 'delete', props: ['beforeEleca'] },
+    { type: 'event', event: 'delete', props: ['afterEleca'] },
+    { type: 'event', event: 'event', props: ['startE', 55, 12, castleAnnounce, 'map0'] },
+    { type: 'event', event: 'event', props: ['startE', 55, 13, castleAnnounce, 'map0'] },
+    {
+      type: 'event',
+      event: 'kill',
+      props: [
+        [28, 45],
+        [29, 44],
+        [30, 45],
+        [24, 28],
+        [25, 28],
+      ],
+    },
+    { type: 'endTimeline' },
+  ],
+};
