+++ src/scenes/Map3.ts
@@ -2,6 +2,7 @@ import { sceneKeys } from './sceneKeys';
 import { json, system } from 'index';
 import { Map_TPL } from './Map.tpl';
 import Mel from '@/assets/characters/static/boss/melcine.png';
+import { warp0 } from 'classes/timelineWords';
 import {
   restart0,
   restart1,
@@ -18,8 +19,7 @@ import {
   goMelcine,
   beforeMelcineBattle,
   afterMelcineBattle,
-  warp0,
-} from 'classes/timelineWords';
+} from 'classes/timelineWords3';
 import { pouler, shiden } from 'friends';
 
 export class Map3 extends Map_TPL {
