+++ src/scenes/Map5.ts
@@ -6,24 +6,7 @@ import Bte from '@/assets/characters/static/boss/bte.png';
 import Mel from '@/assets/characters/static/boss/melcine.png';
 import Eleca from '@/assets/characters/static/boss/eleca.png';
 import Obc from '@/assets/characters/static/boss/obc.png';
-import {
-  restart0,
-  restart1,
-  restart2,
-  II,
-  III,
-  IV,
-  reI,
-  reII,
-  reIII,
-  warpstart,
-  explanation3,
-  addPouler,
-  goMelcine,
-  beforeMelcineBattle,
-  afterMelcineBattle,
-  warp0,
-} from 'classes/timelineWords';
+import { warp0 } from 'classes/timelineWords';
 import { funcs } from 'classes/exam';
 import { pouler, shiden } from 'friends';
 
@@ -45,21 +28,6 @@ export class Map5 extends Map_TPL {
     super.create();
     this.setBoss(51, 24, 'Ate');
     super.setEvent('exit', warp0);
-    super.setHint('explanation2', explanation3);
-    super.setEvent('add3', addPouler);
-    super.setEvent('goMelcine', goMelcine);
-    super.setEvent('beforeMelcine', beforeMelcineBattle);
-    super.setEvent('afterMelcine', afterMelcineBattle);
-    super.setEvent('restart', restart0);
-    super.setEvent('restart1', restart1);
-    super.setEvent('restart2', restart2);
-    super.setHint('woop', II);
-    super.setHint('woop1', III);
-    super.setHint('woop2', IV);
-    super.setHint('woopto1', reI);
-    super.setHint('woop1to2', reII);
-    super.setHint('wooptoboss', reIII);
-    super.setHint('woopstart', warpstart);
     funcs.set(this.name + ',create', (s: any[]) => {
       if (s[0] == 'Bte') this.setBoss(0, 0, 'Bte');
       if (s[0] == 'Mel') this.setBoss(0, 0, 'Mel');
